we can add only 10 pizzas, else Exception countOrderPrice() added countDiscount() added useDiscount() added

+++ src/main/java/Domain/Pizza.java
@@ -1,13 +1,15 @@
 package Domain;
 
+import java.math.BigDecimal;
+
 public class Pizza {
 
     private Long id;
     private String name;
-    private Double price;
+    private BigDecimal price;
     private PizzaType pizzaType;
 
-    public Pizza(Long id, String name, Double price, PizzaType pizzaType) {
+    public Pizza(Long id, String name, BigDecimal price, PizzaType pizzaType) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -40,11 +42,11 @@ public class Pizza {
         this.name = name;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
